Related to Issue #13
Made the necessary changes. the App.text. Additionally, there is support for deadlines and timeouts for ports. The devices are also specifically identified in the components. The device managers tell which device they are implementing. An application can request devices with specific tags. For example, in the topology if there are four PMUs, north, south, east and west,; each PMU is connected to one BBB. An application actor can indicate that it's x component needs to be collocated with the north PMU. If not specified, an actor can be deployed on any node.

Look at the weather monitor example.